feat: improve search box UX and accessibility
