+++ src/components/features/landing/hero/Hero.tsx
@@ -4,15 +4,24 @@ import { SearchBox } from '@/components/ui/composites/search-box/SearchBox';
 import { ContentContainer } from '@/components/ui/content-container';
 import { SectionContainer } from '@/components/ui/section-container';
 
-import { SearchPreview } from '../search-preview/SearchPreview';
+import { SearchPreview } from '../../../ui/composites/search-preview/SearchPreview';
 import { ChainStats } from './chain-stats/ChainStats';
 import { useHero } from './hook';
 
 export interface HeroProps {}
 
 export const Hero = (props: HeroProps) => {
-  const { handleKeywordChange, searchError, searchLoading, searchResult } =
-    useHero();
+  const {
+    handleKeywordChange,
+    handleInputFocus,
+    handleKeyDown,
+    isResultVisible,
+    resultRef,
+    inputRef,
+    searchError,
+    searchLoading,
+    searchResult
+  } = useHero();
 
   return (
     <SectionContainer>
@@ -30,15 +39,22 @@ export const Hero = (props: HeroProps) => {
 
         <div className="relative mx-auto w-full max-w-3xl">
           <SearchBox
+            ref={inputRef}
+            onFocus={handleInputFocus}
+            onKeyDown={handleKeyDown}
             placeholder="Search transaction hash, account id, or block number/hash"
             onKeywordChange={handleKeywordChange}
           />
 
-          <SearchPreview
-            searchResult={searchResult}
-            isLoading={searchLoading}
-            error={searchError}
-          />
+          {isResultVisible && (
+            <SearchPreview
+              ref={resultRef}
+              onKeyDown={handleKeyDown}
+              searchResult={searchResult}
+              isLoading={searchLoading}
+              error={searchError}
+            />
+          )}
         </div>
 
         <ChainStats />
