+++ src/components/features/landing/hero/hook.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import { toast } from 'sonner';
+import { useOnClickOutside } from 'usehooks-ts';
 
 import api from '@/api';
 import type { SearchAllResponse } from '@/schemas/searchs';
@@ -10,6 +11,22 @@ export const useHero = () => {
   const [searchResult, setSearchResult] = useState<SearchAllResponse>();
   const [searchLoading, setSearchLoading] = useState(false);
   const [searchError, setSearchError] = useState<string>();
+  const [isResultVisible, setIsResultVisible] = useState(false);
+
+  const inputRef = useRef<HTMLDivElement>(null);
+  const resultRef = useRef<HTMLDivElement>(null);
+  useOnClickOutside([resultRef, inputRef], () => setIsResultVisible(false));
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLDivElement>) => {
+    if (e.key === 'Escape') {
+      e.currentTarget.blur();
+      setIsResultVisible(false);
+    }
+  };
+
+  const handleInputFocus = () => {
+    setIsResultVisible(true);
+  };
 
   const handleKeywordChange = async (val: string) => {
     const keyword = val.trim();
@@ -34,7 +51,12 @@ export const useHero = () => {
   };
 
   return {
+    isResultVisible,
+    resultRef,
+    inputRef,
     handleKeywordChange,
+    handleKeyDown,
+    handleInputFocus,
     searchResult,
     searchLoading,
     searchError
